[#61] refactor: dev 병합 후 충돌 해결
