+++ admin/src/main/java/itcast/application/AdminBlogService.java
@@ -8,9 +8,9 @@ import itcast.dto.request.AdminBlogRequest;
 import itcast.dto.response.AdminBlogResponse;
 import itcast.exception.IdNotFoundException;
 import itcast.exception.NotAdminException;
+import itcast.jwt.repository.UserRepository;
 import itcast.repository.AdminRepository;
 import itcast.repository.BlogRepository;
-import itcast.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
