+++ admin/src/main/java/itcast/application/AdminNewsService.java
@@ -7,10 +7,10 @@ import itcast.dto.request.AdminNewsRequest;
 import itcast.dto.response.AdminNewsResponse;
 import itcast.exception.IdNotFoundException;
 import itcast.exception.NotAdminException;
+import itcast.jwt.repository.UserRepository;
 import itcast.repository.AdminRepository;
 import itcast.repository.NewsRepository;
 import java.time.LocalDate;
-import itcast.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
