+++ admin/src/main/java/itcast/controller/AdminNewsController.java
@@ -1,14 +1,15 @@
 package itcast.controller;
 
 import itcast.application.AdminNewsService;
-import itcast.auth.jwt.CheckAuth;
-import itcast.auth.jwt.LoginMember;
 import itcast.domain.news.enums.NewsStatus;
 import itcast.dto.request.AdminNewsRequest;
 import itcast.dto.response.AdminNewsResponse;
 import itcast.ResponseTemplate;
 import itcast.dto.response.PageResponse;
 import java.time.LocalDate;
+
+import itcast.jwt.CheckAuth;
+import itcast.jwt.LoginMember;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
