+++ admin/src/test/java/itcast/AdminBlogServiceTest.java
@@ -8,9 +8,9 @@ import itcast.domain.user.User;
 import itcast.domain.user.enums.Interest;
 import itcast.dto.request.AdminBlogRequest;
 import itcast.dto.response.AdminBlogResponse;
+import itcast.jwt.repository.UserRepository;
 import itcast.repository.AdminRepository;
 import itcast.repository.BlogRepository;
-import itcast.user.repository.UserRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
