+++ admin/src/test/java/itcast/AdminNewsServiceTest.java
@@ -7,9 +7,9 @@ import itcast.domain.user.User;
 import itcast.domain.user.enums.Interest;
 import itcast.dto.request.AdminNewsRequest;
 import itcast.dto.response.AdminNewsResponse;
+import itcast.jwt.repository.UserRepository;
 import itcast.repository.AdminRepository;
 import itcast.repository.NewsRepository;
-import itcast.user.repository.UserRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
